p4i integration Integrate 208580 from 4.1 to main: Fixes:    compiler warnings
[git-p4: depot-paths = "//depot/qt/main/": change = 208581]

+++ src/gui/painting/qstroker.cpp
@@ -766,7 +766,7 @@ QPointF qt_curves_for_arc(const QRectF &rect, qreal startAngle, qreal sweepLengt
     qreal b = rect.height() / 2.0;
 
     qreal absSweepLength = (sweepLength < 0 ? -sweepLength : sweepLength);
-    int iterations = ceil((absSweepLength) / 90.0);
+    int iterations = (int)ceil((absSweepLength) / 90.0);
 
     QPointF first_point;
